@@ -19,7 +19,7 @@ namespace Exoft.Gamification.Api.Common.Models.User
 
         public int XP { get; set; }
 
-        public int BadgetCount { get; set; }
+        public int BadgesCount { get; set; }
 
         public IEnumerable<string> Roles { get; set; }
 
