@@ -23,10 +23,10 @@ namespace Exoft.Gamification.Api.Data.Core.Entities
 
         public string Status { get; set; }
 
-        public Guid? AvatarId { get; set; }
-        
         public int XP { get; set; }
 
+        public Guid? AvatarId { get; set; }
+        
         public ICollection<UserRoles> Roles { get; set; }
 
         public ICollection<UserAchievement> Achievements { get; set; }
