@@ -1,10 +1,12 @@
-﻿using Exoft.Gamification.Api.Data.Core.Entities;
-using Exoft.Gamification.Api.Data.Core.Helpers;
-using Exoft.Gamification.Api.Data.Core.Interfaces.Repositories;
-using Microsoft.EntityFrameworkCore;
-using System;
+﻿using System;
 using System.Linq;
 using System.Threading.Tasks;
+
+using Exoft.Gamification.Api.Data.Core.Entities;
+using Exoft.Gamification.Api.Data.Core.Helpers;
+using Exoft.Gamification.Api.Data.Core.Interfaces.Repositories;
+
+using Microsoft.EntityFrameworkCore;
 
 namespace Exoft.Gamification.Api.Data.Repositories
 {
@@ -14,11 +16,42 @@ namespace Exoft.Gamification.Api.Data.Repositories
         {
         }
 
-        public async Task<ReturnPagingInfo<UserAchievement>> GetAllAchievementsByUserAsync(PagingInfo pagingInfo, Guid UserId)
+        public async Task<ReturnPagingInfo<UserAchievement>> GetAllAchievementsByUserAsync(PagingInfo pagingInfo, Guid userId)
+        {
+            var take = pagingInfo.PageSize;
+            var skip = (pagingInfo.CurrentPage - 1) * pagingInfo.PageSize;
+
+            var query = IncludeAll()
+                .Where(i => i.User.Id == userId)
+                .OrderByDescending(s => s.AddedTime)
+                .Select(i => new
+                {
+                    Data = i,
+                    TotalCount = IncludeAll().Count()
+                });
+
+            var entities = pagingInfo.PageSize != 0
+                               ? await query.Skip(skip).Take(take).ToListAsync()
+                               : await query.ToListAsync();
+
+            var totalCount = entities.First().TotalCount;
+
+            var result = new ReturnPagingInfo<UserAchievement>
+            {
+                CurrentPage = pagingInfo.CurrentPage,
+                PageSize = entities.Count,
+                TotalItems = totalCount,
+                TotalPages = (int)Math.Ceiling((double)totalCount / (pagingInfo.PageSize == 0 ? totalCount : pagingInfo.PageSize)),
+                Data = entities.Select(i => i.Data).ToList()
+            };
+
+            return result;
+        }
+
+        public async Task<ReturnPagingInfo<UserAchievement>> GetAllUsersByAchievementAsync(PagingInfo pagingInfo, Guid achievementId)
         {
             var query = IncludeAll()
-                .Where(o => o.User.Id == UserId)
-                .OrderByDescending(i => i.AddedTime)
+                .Where(o => o.Achievement.Id == achievementId)
                 .AsQueryable();
 
             if (pagingInfo.PageSize != 0)
@@ -28,6 +61,16 @@ namespace Exoft.Gamification.Api.Data.Repositories
             }
 
             var items = await query.ToListAsync();
+
+            //// todo: use this code after update core >= 3.0.0
+            ////var result = new ReturnPagingInfo<UserAchievement>()
+            ////{
+            ////    CurrentPage = pagingInfo.CurrentPage,
+            ////    PageSize = items.Count,
+            ////    TotalItems = query.Count(),
+            ////    TotalPages = (int)Math.Ceiling((double)query.Count() / (pagingInfo.PageSize == 0 ? query.Count() : pagingInfo.PageSize)),
+            ////    Data = items
+            ////};
 
             int itemsCount = await query.CountAsync();
 
@@ -53,13 +96,6 @@ namespace Exoft.Gamification.Api.Data.Repositories
         {
             return await Context.UserAchievement
                 .CountAsync(o => o.User.Id == userId);
-        }
-
-        public async Task<int> GetSummaryXpByUserAsync(Guid userId)
-        {
-            return await IncludeAll()
-                .Where(o => o.User.Id == userId)
-                .SumAsync(i => i.Achievement.XP);
         }
 
         protected override IQueryable<UserAchievement> IncludeAll()
