@@ -1,11 +1,12 @@
-﻿using Exoft.Gamification.Api.Common.Models;
-using System;
+﻿using System;
 using System.Threading.Tasks;
+
+using Exoft.Gamification.Api.Common.Models;
 
 namespace Exoft.Gamification.Api.Services.Interfaces.Services
 {
     public interface IFileService
     {
-        Task<FileModel> GetFileByIdAsync(Guid Id);
+        Task<FileModel> GetFileByIdAsync(Guid id);
     }
 }
