@@ -1,12 +1,15 @@
-﻿using Exoft.Gamification.Api.Common.Models;
+﻿using System;
+using System.Threading.Tasks;
+
+using Exoft.Gamification.Api.Common.Models.RequestAchievement;
 using Exoft.Gamification.Api.Data.Core.Helpers;
 using Exoft.Gamification.Api.Services.Interfaces.Services;
+
 using FluentValidation;
 using FluentValidation.AspNetCore;
+
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using System;
-using System.Threading.Tasks;
 
 namespace Exoft.Gamification.Api.Controllers
 {
@@ -15,20 +18,17 @@ namespace Exoft.Gamification.Api.Controllers
     [ApiController]
     public class RequestAchievementController : GamificationController
     {
-        private readonly IAchievementService _achievementService;
         private readonly IRequestAchievementService _requestAchievementService;
-        private readonly IValidator<RequestAchievementModel> _requestAchievementModelValidator;
+        private readonly IValidator<CreateRequestAchievementModel> _createRequestAchievementModelValidator;
 
         public RequestAchievementController
         (
-            IAchievementService achievementService,
             IRequestAchievementService requestAchievementService,
-            IValidator<RequestAchievementModel> requestAchievementModelValidator
+            IValidator<CreateRequestAchievementModel> createRequestAchievementModelValidator
         )
         {
-            _achievementService = achievementService;
             _requestAchievementService = requestAchievementService;
-            _requestAchievementModelValidator = requestAchievementModelValidator;
+            _createRequestAchievementModelValidator = createRequestAchievementModelValidator;
         }
 
         /// <summary>
@@ -37,9 +37,9 @@ namespace Exoft.Gamification.Api.Controllers
         /// <response code="200">When request success sended and added</response>
         /// <response code="422">When the model structure is correct but validation fails</response>
         [HttpPost]
-        public async Task<IActionResult> AddRequestAsync([FromBody] RequestAchievementModel model)
+        public async Task<IActionResult> AddRequestAsync([FromBody] CreateRequestAchievementModel model)
         {
-            var resultValidation = await _requestAchievementModelValidator.ValidateAsync(model);
+            var resultValidation = await _createRequestAchievementModelValidator.ValidateAsync(model);
             resultValidation.AddToModelState(ModelState, null);
 
             if (!ModelState.IsValid)
@@ -53,8 +53,10 @@ namespace Exoft.Gamification.Api.Controllers
         }
 
         /// <summary>
-        /// Returns all achievement requests 
+        /// Return all achievement requests 
         /// </summary>
+        /// <response code="200">Return all requests</response>
+        /// <response code="403">When user don't have permissions to this action</response>
         [HttpGet]
         [Authorize(Roles = GamificationRole.Admin)]
         public async Task<IActionResult> GetAllAchievementRequests()
@@ -63,8 +65,11 @@ namespace Exoft.Gamification.Api.Controllers
         }
 
         /// <summary>
-        /// Deletes Request 
+        /// Delete request 
         /// </summary>
+        /// <response code="200">Decline achievement request with current Id</response>
+        /// <response code="403">When user don't have permissions to this action</response>
+        /// <response code="404">When request with current Id is not found</response>
         [HttpDelete("{id}")]
         [Authorize(Roles = GamificationRole.Admin)]
         public async Task<IActionResult> DeclineRequest(Guid id)
@@ -79,14 +84,22 @@ namespace Exoft.Gamification.Api.Controllers
             return Ok();
         }
 
-
         /// <summary>
-        /// Approves Request  
+        /// Approve request  
         /// </summary>
+        /// <response code="200">When request is approved</response>
+        /// <response code="403">When user don't have permissions to this action</response>
+        /// <response code="404">When request with current Id is not found</response>
         [HttpPost("{id}")]
         [Authorize(Roles = GamificationRole.Admin)]
         public async Task<IActionResult> ApproveRequest(Guid id)
         {
+            var achievementRequest = await _requestAchievementService.GetByIdAsync(id);
+            if (achievementRequest == null)
+            {
+                return NotFound();
+            }
+
             await _requestAchievementService.ApproveAchievementRequestAsync(id);
 
             return Ok();
