@@ -1,10 +1,13 @@
-﻿using Exoft.Gamification.Api.Common.Models.Thank;
+﻿using System.Threading.Tasks;
+
+using Exoft.Gamification.Api.Common.Models.Thank;
 using Exoft.Gamification.Api.Services.Interfaces.Services;
+
 using FluentValidation;
 using FluentValidation.AspNetCore;
+
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using System.Threading.Tasks;
 
 namespace Exoft.Gamification.Api.Controllers
 {
@@ -26,6 +29,11 @@ namespace Exoft.Gamification.Api.Controllers
             _createThankModelValidator = createThankModelValidator;
         }
 
+        /// <summary>
+        /// Create thank for user.
+        /// </summary>
+        /// <responce code="200">When thank successful created</responce> 
+        /// <response code="422">When the model structure is correct but validation fails</response>
         [HttpPost]
         public async Task<IActionResult> SayThanks([FromBody] CreateThankModel model)
         {
@@ -36,11 +44,15 @@ namespace Exoft.Gamification.Api.Controllers
             {
                 return UnprocessableEntity();
             }
-            
+
             await _thankService.AddAsync(model, UserId);
             return Ok();
         }
 
+        /// <summary>
+        /// Get thank for current user.
+        /// </summary>
+        /// <responce code="200">Thank model</responce> 
         [HttpGet]
         public async Task<IActionResult> GetThanks()
         {
