@@ -1,10 +1,15 @@
-﻿using Exoft.Gamification.Api.Common.Models;
+﻿using System;
+using System.Threading.Tasks;
+
+using Exoft.Gamification.Api.Common.Models;
 using Exoft.Gamification.Api.Data.Core.Helpers;
 using Exoft.Gamification.Api.Services.Interfaces.Services;
+
+using FluentValidation;
+using FluentValidation.AspNetCore;
+
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using System;
-using System.Threading.Tasks;
 
 namespace Exoft.Gamification.Api.Controllers
 {
@@ -15,22 +20,27 @@ namespace Exoft.Gamification.Api.Controllers
     {
         private readonly IUserService _userService;
         private readonly IUserAchievementService _userAchievementService;
+        private readonly IValidator<PagingInfo> _pagingInfoValidator;
 
         public UserAchievementController
         (
             IUserService userService,
-            IUserAchievementService userAchievementsService
+            IUserAchievementService userAchievementsService,
+            IValidator<PagingInfo> pagingInfoValidator
         )
         {
             _userService = userService;
             _userAchievementService = userAchievementsService;
+            _pagingInfoValidator = pagingInfoValidator;
         }
 
         /// <summary>
         /// Add or remove achievements into user
         /// </summary>
-        /// <responce code="200">Return Ok</responce> 
-        /// <responce code="404">When any achievement or user does not exist</responce> 
+        /// <responce code="200">Return Ok</responce>
+        /// <response code="403">When user don't have permissions to this action</response>
+        /// <responce code="404">When any achievement or user does not exist</responce>
+        /// <response code="422">When the model structure is correct but validation fails</response>
         [Authorize(Roles = GamificationRole.Admin)]
         [HttpPost("{userId}/achievements/")]
         public async Task<IActionResult> UpdateUserAchievements([FromBody] AssignAchievementsToUserModel model, Guid userId)
@@ -49,10 +59,19 @@ namespace Exoft.Gamification.Api.Controllers
         /// <summary>
         /// Get paged list of achievements current user
         /// </summary>
-        /// <responce code="200">Return all achievements current user</responce> 
+        /// <responce code="200">Return all achievements current user</responce>
+        /// <response code="422">When the model structure is correct but validation fails</response>
         [HttpGet("{userId}/achievements")]
         public async Task<IActionResult> GetUserAchievementsAsync(Guid userId, [FromQuery] PagingInfo pagingInfo)
         {
+            var resultValidation = await _pagingInfoValidator.ValidateAsync(pagingInfo);
+            resultValidation.AddToModelState(ModelState, null);
+
+            if (!ModelState.IsValid)
+            {
+                return UnprocessableEntity(ModelState);
+            }
+
             var item = await _userAchievementService.GetAllAchievementsByUserAsync(pagingInfo, userId);
 
             return Ok(item);
